+++ src/thor/timedistancebssmatrix.cc
@@ -164,7 +164,8 @@ std::vector<TimeDistance> TimeDistanceBSSMatrix::OneToMany(
     GraphReader& graphreader,
     const sif::mode_costing_t& mode_costing,
     const TravelMode /*mode*/,
-    const float max_matrix_distance) {
+    const float max_matrix_distance,
+    const uint32_t matrix_locations) {
 
   pedestrian_costing_ = mode_costing[static_cast<uint32_t>(TravelMode::kPedestrian)];
   bicycle_costing_ = mode_costing[static_cast<uint32_t>(TravelMode::kBicycle)];
@@ -227,7 +228,8 @@ std::vector<TimeDistance> TimeDistanceBSSMatrix::OneToMany(
       // have been settled.
       tile = graphreader.GetGraphTile(pred.edgeid());
       const DirectedEdge* edge = tile->directededge(pred.edgeid());
-      if (UpdateDestinations(origin, locations, destedge->second, edge, tile, pred)) {
+      if (UpdateDestinations(origin, locations, destedge->second, edge, tile, pred,
+                             matrix_locations)) {
         return FormTimeDistanceMatrix();
       }
     }
@@ -358,7 +360,8 @@ std::vector<TimeDistance> TimeDistanceBSSMatrix::ManyToOne(
     GraphReader& graphreader,
     const sif::mode_costing_t& mode_costing,
     const TravelMode /*mode*/,
-    const float max_matrix_distance) {
+    const float max_matrix_distance,
+    const uint32_t matrix_locations) {
 
   pedestrian_costing_ = mode_costing[static_cast<uint32_t>(TravelMode::kPedestrian)];
   bicycle_costing_ = mode_costing[static_cast<uint32_t>(TravelMode::kBicycle)];
@@ -415,7 +418,7 @@ std::vector<TimeDistance> TimeDistanceBSSMatrix::ManyToOne(
       // have been settled.
       tile = graphreader.GetGraphTile(pred.edgeid());
       const DirectedEdge* edge = tile->directededge(pred.edgeid());
-      if (UpdateDestinations(dest, locations, destedge->second, edge, tile, pred)) {
+      if (UpdateDestinations(dest, locations, destedge->second, edge, tile, pred, matrix_locations)) {
         return FormTimeDistanceMatrix();
       }
     }
@@ -448,20 +451,23 @@ std::vector<TimeDistance> TimeDistanceBSSMatrix::SourceToTarget(
     baldr::GraphReader& graphreader,
     const sif::mode_costing_t& mode_costing,
     const sif::TravelMode _,
-    const float max_matrix_distance) {
+    const float max_matrix_distance,
+    const uint32_t matrix_locations) {
   // Run a series of one to many calls and concatenate the results.
   std::vector<TimeDistance> many_to_many;
   if (source_location_list.size() <= target_location_list.size()) {
     for (const auto& origin : source_location_list) {
       std::vector<TimeDistance> td =
-          OneToMany(origin, target_location_list, graphreader, mode_costing, _, max_matrix_distance);
+          OneToMany(origin, target_location_list, graphreader, mode_costing, _, max_matrix_distance,
+                    matrix_locations);
       many_to_many.insert(many_to_many.end(), td.begin(), td.end());
       Clear();
     }
   } else {
     for (const auto& destination : target_location_list) {
-      std::vector<TimeDistance> td = ManyToOne(destination, source_location_list, graphreader,
-                                               mode_costing, _, max_matrix_distance);
+      std::vector<TimeDistance> td =
+          ManyToOne(destination, source_location_list, graphreader, mode_costing, _,
+                    max_matrix_distance, matrix_locations);
       many_to_many.insert(many_to_many.end(), td.begin(), td.end());
       Clear();
     }
@@ -696,7 +702,8 @@ bool TimeDistanceBSSMatrix::UpdateDestinations(
     std::vector<uint32_t>& destinations,
     const DirectedEdge* edge,
     const graph_tile_ptr& tile,
-    const EdgeLabel& pred) {
+    const EdgeLabel& pred,
+    const uint32_t matrix_locations) {
   // For each destination along this edge
   for (auto dest_idx : destinations) {
     Destination& dest = destinations_[dest_idx];
@@ -777,7 +784,10 @@ bool TimeDistanceBSSMatrix::UpdateDestinations(
   if (allfound) {
     current_cost_threshold_ = maxcost;
   }
-  return settled_count_ == destinations_.size();
+
+  // Return true if the settled count equals the number of destinations or
+  // exceeds the matrix location count provided.
+  return settled_count_ == destinations_.size() || settled_count_ >= matrix_locations;
 }
 
 // Form the time, distance matrix from the destinations list