+++ valhalla/thor/timedistancebssmatrix.h
@@ -42,6 +42,9 @@ public:
    * @param  mode          Travel mode to use. Actually It doesn't make sense in matrix_bss, because
    * the travel mode must be pedestrian and bicycle
    * @param  max_matrix_distance   Maximum arc-length distance for current mode.
+   * @param  matrix_locations      Number of matrix locations to satisfy a one to many or many to
+   *                               one request. This allows partial results: e.g. find time/distance
+   *                               to the closest 20 out of 50 locations).
    * @return time/distance from origin index to all other locations
    */
   std::vector<TimeDistance>
@@ -50,7 +53,8 @@ public:
             baldr::GraphReader& graphreader,
             const sif::mode_costing_t& mode_costing,
             const sif::TravelMode /*mode*/,
-            const float max_matrix_distance);
+            const float max_matrix_distance,
+            const uint32_t matrix_locations = kAllLocations);
 
   /**
    * Many to one time and distance cost matrix. Computes time and distance
@@ -62,6 +66,9 @@ public:
    * @param  mode          Travel mode to use. Actually It doesn't make sense in matrix_bss, because
    * the travel mode must be pedestrian and bicycle
    * @param  max_matrix_distance   Maximum arc-length distance for current mode.
+   * @param  matrix_locations      Number of matrix locations to satisfy a one to many or many to
+   *                               one request. This allows partial results: e.g. find time/distance
+   *                               to the closest 20 out of 50 locations).
    * @return time/distance to the destination index from all other locations
    */
   std::vector<TimeDistance>
@@ -70,7 +77,8 @@ public:
             baldr::GraphReader& graphreader,
             const sif::mode_costing_t& mode_costing,
             const sif::TravelMode /*mode*/,
-            const float max_matrix_distance);
+            const float max_matrix_distance,
+            const uint32_t matrix_locations = kAllLocations);
 
   /**
    * Many to many time and distance cost matrix. Computes time and distance
@@ -100,6 +108,9 @@ public:
    * @param  mode                  Travel mode to use. Actually It doesn't make sense in matrix_bss,
    * because the travel mode must be pedestrian and bicycle
    * @param  max_matrix_distance   Maximum arc-length distance for current mode.
+   * @param  matrix_locations      Number of matrix locations to satisfy a one to many or many to
+   *                               one request. This allows partial results: e.g. find time/distance
+   *                               to the closest 20 out of 50 locations).
    * @return time/distance from origin index to all other locations
    */
   std::vector<TimeDistance>
@@ -108,7 +119,8 @@ public:
                  baldr::GraphReader& graphreader,
                  const sif::mode_costing_t& mode_costing,
                  const sif::TravelMode /*mode*/,
-                 const float max_matrix_distance);
+                 const float max_matrix_distance,
+                 const uint32_t matrix_locations = kAllLocations);
 
   /**
    * Clear the temporary information generated during time+distance
@@ -241,6 +253,10 @@ protected:
    * @param   edge          Directed edge
    * @param   pred          Predecessor information in shortest path.
    * @param   predindex     Predecessor index in EdgeLabels vector.
+   * @param   matrix_locations Count of locations that must be found. When provided it allows
+   *                           a partial result to be returned (e.g. best 20 out of 50 locations).
+   *                           When not supplied in the request this is set to max uint32_t value
+   *                           so that all supplied locations must be settled.
    * @return  Returns true if all destinations have been settled.
    */
   bool UpdateDestinations(const valhalla::Location& origin,
@@ -248,7 +264,8 @@ protected:
                           std::vector<uint32_t>& destinations,
                           const baldr::DirectedEdge* edge,
                           const graph_tile_ptr& tile,
-                          const sif::EdgeLabel& pred);
+                          const sif::EdgeLabel& pred,
+                          const uint32_t matrix_locations);
 
   /**
    * Form a time/distance matrix from the results.